requirements_tmp.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.104.0
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.2.3
h11==0.14.0
idna==3.4
psycopg2==2.9.10
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.10.1
redis==5.0.1
smmap==5.0.2
sniffio==1.3.0
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.27.0
types-pyOpenSSL==23.3.0.0
types-redis==4.6.0.7
typing_extensions==4.8.0
uvicorn==0.23.2
